GCPRofiler: pushing the handling to the thread level

+++ runtime/gc_profiler/MProfiler.cc
@@ -2680,7 +2680,7 @@ void GCHistogramThreadManager::initHistograms(void){
 	memset((void*)lastWindowHistTable, 0, lastWindowHistSize);
 
 
-	for(size_t i = 0; i < kGCMMPMaxHistogramEntries; i++){
+	for(int i = 0; i < kGCMMPMaxHistogramEntries; i++){
 		histogramTable[i].index 			= (i+1) * 1.0;
 		lastWindowHistTable[i].index  = (i+1) * 1.0;
 	}
